Started making getAttendance() method.
The method getAttendance is written and should work (Possibly), but I am
unable to make a box to print the results of the method to.

Look into making an uneditable box wherever you wish in the GUI and try
and see if the method works with it. This box should store the result of
the getAttendance() method.