Fixed potential bug for max length strings in database classs
In the database class, the SQL statements that are formed weren't compensating for the null terminator on the end of the strings we were passing in. These weren't a problem during testing because we weren't using max lengths for each string we passed in. However, if we did, we'd have stack corruption